Removing the need for the user creds on scan
Later on we will need to add this back. We have about ~633 users from
the previous iteration of the bot that I would like to re-introduce. In
order to do that we need to not have their token and only use the bot's
token. For now, we do this. Later on when we need their creds for
something we can put a message in their channel to re-authenticate.