Making scans multi-threaded
We will try this and if it fails horribly we can roll this commit back.
It looks like the twitch api is very well behaved but the combination of
EF Core and MySQL are behaving very poorly. I think I've worked through
enough of the issues to move forward but we will see how it goes.