Adding storing twitch bans and messages
This is a very hamfisted way of doing this but we need to collect the
logs sooner than later to re-baseline what the current scam messages
look like.

To elaborate, the reason we are doing this is not because we're nosy.
It's so that we can cross reference the bans with specific chat messages
so that we can understand what scam messages bots are typing. For
example, when a bot comes in and types a first time message with
something like, "Want more viewers? scam.com" Other than that I couldn't
care less what's in the chat logs.