fix: race condition when setting bot token and ignoring banned users
There is still another race condition but this should severly limit the
surface area of the issue. We need to centralize where the bot token is
refreshed instead of allowing 2 separate threads to compete for it.

We always wanted to ignore banned users but I forget to exclude them
from the mod list.