fix: concurrency issue when saving records
I don't have high hopes for this fix but it was suggested here and I
can't completely ignore the similiarity in how odd this is:
https://learn.microsoft.com/en-us/answers/questions/1186074/database-update-insert-fails
So we shouldn't have the possibility of concurrency issues here because
we are simply inserting everything. Worse case scenario we would end up
with double entries for the same ban. Yet for some reason it seems to
think we're modifying entries when we're supposed to be inserting them.
As such it's reasonable to assume swapping the state to added (despite
the fact that the "AddRange" method is supposed to do that for us) might
resolve the issue. Again, I'm not entirely hopeful but I'm willing to
try. Since the issue is intermittent we have to test in prod.

nullinside-development-group/nullinside-api#70