feat: only joining registered user channels
We joined all channels we were a mod in just to be nice to users that
haven't touched the site in a while and may have fallen out of the
database. The result is that startup has gotten quite long as the bot
has to join hundreds of channels. This code will allow us to join
channels as they go live resulting in a faster startup.